Refactor Footer: Pull out Navigation components
The next few commits will involve refactoring the footer. Instead
of having all of the components in one file, we are going to split
up the components into multiple files, then import what we need
to create the footer itself.
